moved the BddLibConstants to constants package
