+++ bdd-cucumber-gherkin-lib-core/src/main/kotlin/com/ragin/bdd/cucumber/constants/BddLibConstants.kt
@@ -1,4 +1,4 @@
-package com.ragin.bdd.cucumber
+package com.ragin.bdd.cucumber.constants
 
 import java.util.regex.Pattern
